requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
mysqlclient>=2.2.0
python-decouple>=3.8
PyJWT>=2.8.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-dateutil>=2.8.0
